move ledger restoration to AnvilAlert
