+++ src/components/lend/LendAnvil.tsx
@@ -81,29 +81,28 @@ export const LendAnvil: React.FC<ReserveComponentProps> = ({ poolId, assetId }) 
   }
 
   // verify that the user can act
-  const { isSubmitDisabled, isMaxDisabled, reason, disabledType, isError, requiresRestore } =
-    useMemo(
-      () =>
-        getErrorFromSim(simResponse, () => {
-          const errorProps: Partial<SubmitError> = {};
-          if (!toLend) {
-            errorProps.isSubmitDisabled = true;
-            errorProps.isError = true;
-            errorProps.isMaxDisabled = false;
-            errorProps.reason = 'Please enter an amount to lend.';
-            errorProps.disabledType = 'info';
-          } else if (toLend.split('.')[1]?.length > decimals) {
-            setValidDecimals(false);
-            errorProps.isSubmitDisabled = true;
-            errorProps.isError = true;
-            errorProps.isMaxDisabled = false;
-            errorProps.reason = `You cannot input more than ${decimals} decimal places.`;
-            errorProps.disabledType = 'warning';
-          }
-          return errorProps;
-        }),
-      [freeUserBalanceScaled, toLend, simResponse]
-    );
+  const { isSubmitDisabled, isMaxDisabled, reason, disabledType, isError } = useMemo(
+    () =>
+      getErrorFromSim(simResponse, () => {
+        const errorProps: Partial<SubmitError> = {};
+        if (!toLend) {
+          errorProps.isSubmitDisabled = true;
+          errorProps.isError = true;
+          errorProps.isMaxDisabled = false;
+          errorProps.reason = 'Please enter an amount to lend.';
+          errorProps.disabledType = 'info';
+        } else if (toLend.split('.')[1]?.length > decimals) {
+          setValidDecimals(false);
+          errorProps.isSubmitDisabled = true;
+          errorProps.isError = true;
+          errorProps.isMaxDisabled = false;
+          errorProps.reason = `You cannot input more than ${decimals} decimal places.`;
+          errorProps.disabledType = 'warning';
+        }
+        return errorProps;
+      }),
+    [freeUserBalanceScaled, toLend, simResponse]
+  );
   const handleLendMax = () => {
     if (userPoolData) {
       if (freeUserBalanceScaled > 0) {
@@ -183,7 +182,7 @@ export const LendAnvil: React.FC<ReserveComponentProps> = ({ poolId, assetId }) 
           </Box>
         </Box>
         {!isError && (
-          <TxOverview simResponse={simResponse} requiresRestore={requiresRestore}>
+          <TxOverview>
             {!isLoading && (
               <>
                 <Value title="Amount to supply" value={`${toLend ?? '0'} ${symbol}`} />
@@ -224,7 +223,9 @@ export const LendAnvil: React.FC<ReserveComponentProps> = ({ poolId, assetId }) 
             )}
           </TxOverview>
         )}
-        {isError && <AnvilAlert severity={disabledType} message={reason} />}
+        {isError && (
+          <AnvilAlert severity={disabledType} message={reason} simResponse={simResponse} />
+        )}
       </Section>
     </Row>
   );
