+++ src/components/withdraw/WithdrawAnvil.tsx
@@ -84,15 +84,7 @@ export const WithdrawAnvil: React.FC<ReserveComponentProps> = ({ poolId, assetId
     </OpaqueButton>
   );
   // verify that the user can act
-  const {
-    isSubmitDisabled,
-    isMaxDisabled,
-    reason,
-    disabledType,
-    extraContent,
-    isError,
-    requiresRestore,
-  } = useMemo(
+  const { isSubmitDisabled, isMaxDisabled, reason, disabledType, extraContent, isError } = useMemo(
     () =>
       getErrorFromSim(simResponse, () => {
         const errorProps: Partial<SubmitError> = {};
@@ -236,7 +228,7 @@ export const WithdrawAnvil: React.FC<ReserveComponentProps> = ({ poolId, assetId
           </Box>
         </Box>
         {!isError && (
-          <TxOverview simResponse={simResponse} requiresRestore={requiresRestore}>
+          <TxOverview>
             {!isLoading && (
               <>
                 <Value title="Amount to withdraw" value={`${toWithdraw ?? '0'} ${symbol}`} />
@@ -278,7 +270,12 @@ export const WithdrawAnvil: React.FC<ReserveComponentProps> = ({ poolId, assetId
           </TxOverview>
         )}
         {isError && (
-          <AnvilAlert severity={disabledType} message={reason} extraContent={extraContent} />
+          <AnvilAlert
+            severity={disabledType}
+            message={reason}
+            extraContent={extraContent}
+            simResponse={simResponse}
+          />
         )}
       </Section>
     </Row>
